@@ -11,52 +11,52 @@
   },
   "devDependencies": {
     "@sveltejs/vite-plugin-svelte": "^3.1.2",
-    "@tsconfig/svelte": "^5.0.4",
+    "@tsconfig/svelte": "^4.0.1",
     "@types/cors": "^2.8.17",
-    "@types/express": "^4.17.21",
+    "@types/express": "^5.0.1",
     "@types/react": "^18.3.12",
     "@types/react-dom": "^18.3.1",
-    "@vitejs/plugin-react": "^4.3.3",
-    "concurrently": "^9.0.1",
+    "@vitejs/plugin-react": "^4.4.1",
+    "concurrently": "^9.1.2",
     "cors": "^2.8.5",
-    "express": "^4.21.1",
-    "glob": "^11.0.0",
+    "express": "^5.1.0",
+    "glob": "^11.0.1",
     "picocolors": "^1.1.1",
-    "svelte-check": "^4.0.5",
-    "tslib": "^2.8.0",
-    "typescript": "^5.5.4",
-    "vite": "^5.4.11"
+    "svelte-check": "^4.1.6",
+    "tslib": "^2.8.1",
+    "typescript": "^5.8.3",
+    "vite": "^5.4.18"
   },
   "dependencies": {
     "@dagrejs/dagre": "^1.1.4",
-    "@radix-ui/react-dropdown-menu": "^2.1.2",
-    "@radix-ui/react-slot": "^1.1.1",
+    "@radix-ui/react-dropdown-menu": "^2.1.11",
+    "@radix-ui/react-slot": "^1.2.0",
     "@threlte/core": "^7.3.1",
-    "@xyflow/react": "^12.5.4",
-    "@xyflow/svelte": "^0.1.31",
-    "class-variance-authority": "^0.7.0",
+    "@xyflow/react": "^12.6.0",
+    "@xyflow/svelte": "^0.1.37",
+    "class-variance-authority": "^0.7.1",
     "clsx": "^2.1.1",
     "d3-drag": "^3.0.0",
     "d3-force": "^3.0.0",
     "d3-hierarchy": "^3.1.2",
     "d3-quadtree": "^3.0.1",
     "d3-selection": "^3.0.0",
-    "elkjs": "^0.9.3",
+    "elkjs": "^0.10.0",
     "entitree-flex": "^0.4.1",
-    "html-to-image": "1.11.11",
-    "lucide-react": "^0.453.0",
-    "nanoid": "^5.0.7",
-    "puppeteer": "^22.7.1",
+    "html-to-image": "1.11.13",
+    "lucide-react": "^0.503.0",
+    "nanoid": "^5.1.5",
+    "puppeteer": "^24.7.0",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
-    "react-icons": "^5.3.0",
+    "react-icons": "^5.5.0",
     "react-remark": "^2.1.0",
-    "styled-components": "^6.1.13",
+    "styled-components": "^6.1.17",
     "svelte": "^4.2.19",
-    "svelte-feather-icons": "^4.1.0",
-    "tailwind-merge": "^2.5.4",
+    "svelte-feather-icons": "^4.2.0",
+    "tailwind-merge": "^3.2.0",
     "twind": "^0.16.19",
     "webcola": "^3.4.0",
-    "zustand": "5.0.0"
+    "zustand": "5.0.3"
   }
 }
