@@ -2,7 +2,7 @@
   "name": "ui-components",
   "version": "0.1.0",
   "private": true,
-  "packageManager": "pnpm@9.11.0",
+  "packageManager": "pnpm@10.9.0",
   "scripts": {
     "dev": "pnpm run generate-registry && next dev --port 3004",
     "build": "pnpm run generate-registry && next build",
@@ -12,36 +12,36 @@
     "generate-registry": "node ./scripts/generate-registry.js"
   },
   "dependencies": {
-    "@radix-ui/react-dropdown-menu": "^2.1.2",
-    "@radix-ui/react-icons": "^1.3.0",
-    "@radix-ui/react-select": "^2.1.2",
-    "@radix-ui/react-slider": "^1.2.1",
-    "@radix-ui/react-slot": "^1.1.0",
-    "@radix-ui/react-toggle": "^1.1.0",
-    "@radix-ui/react-toggle-group": "^1.1.0",
-    "@vercel/analytics": "^1.3.1",
-    "@xyflow/react": "^12.5.1",
-    "class-variance-authority": "^0.7.0",
+    "@radix-ui/react-dropdown-menu": "^2.1.11",
+    "@radix-ui/react-icons": "^1.3.2",
+    "@radix-ui/react-select": "^2.2.2",
+    "@radix-ui/react-slider": "^1.3.2",
+    "@radix-ui/react-slot": "^1.2.0",
+    "@radix-ui/react-toggle": "^1.1.6",
+    "@radix-ui/react-toggle-group": "^1.1.7",
+    "@vercel/analytics": "^1.5.0",
+    "@xyflow/react": "^12.6.0",
+    "class-variance-authority": "^0.7.1",
     "clsx": "^2.1.1",
-    "date-fns": "^3.6.0",
-    "lucide-react": "^0.453.0",
-    "next": "14.2.15",
+    "date-fns": "^4.1.0",
+    "lucide-react": "^0.503.0",
+    "next": "15.3.1",
     "react": "^18.3.1",
-    "react-day-picker": "^8.10.1",
+    "react-day-picker": "^9.6.7",
     "react-dom": "^18.3.1",
-    "tailwind-merge": "^2.5.4",
+    "tailwind-merge": "^3.2.0",
     "tailwindcss-animate": "^1.0.7"
   },
   "devDependencies": {
     "@types/node": "^22",
     "@types/react": "^18.3.12",
     "@types/react-dom": "^18.3.1",
-    "eslint": "^8.57.1",
-    "eslint-config-next": "14.2.15",
+    "eslint": "^9.25.1",
+    "eslint-config-next": "15.3.1",
     "postcss": "^8",
-    "prettier": "^3.3.3",
-    "prettier-plugin-tailwindcss": "^0.6.8",
+    "prettier": "^3.5.3",
+    "prettier-plugin-tailwindcss": "^0.6.11",
     "tailwindcss": "^3.4.14",
-    "typescript": "^5.5.4"
+    "typescript": "^5.8.3"
   }
 }
