@@ -4,13 +4,23 @@
   "private": true,
   "main": "index.js",
   "license": "MIT",
-  "dependencies": {
-    "eslint-config-next": "^14.2.15",
-    "eslint-config-prettier": "^9.1.0",
-    "eslint-config-turbo": "^2.2.3",
-    "eslint-plugin-react": "7.37.2"
+  "type": "module",
+  "exports": {
+    "./base": "./base.js",
+    "./next-js": "./next.js",
+    "./react-internal": "./react-internal.js"
   },
-  "devDependencies": {
-    "typescript": "^5.5.4"
+  "dependencies": {
+    "@eslint/js": "^9.25.0",
+    "@next/eslint-plugin-next": "^15.3.0",
+    "eslint": "^9.25.0",
+    "eslint-config-prettier": "^10.1.1",
+    "eslint-plugin-only-warn": "^1.1.0",
+    "eslint-plugin-react": "^7.37.4",
+    "eslint-plugin-react-hooks": "^5.2.0",
+    "eslint-plugin-turbo": "^2.5.0",
+    "globals": "^16.0.0",
+    "typescript": "^5.8.2",
+    "typescript-eslint": "^8.30.1"
   }
 }
