@@ -14,18 +14,18 @@
     "clean": "pnpm -r --parallel exec rimraf dist .turbo node_modules"
   },
   "devDependencies": {
-    "@turbo/gen": "^2.2.3",
-    "eslint": "^8.57.1",
+    "@turbo/gen": "^2.5.0",
+    "eslint": "^9.25.1",
     "eslint-config-xyflow": "workspace:*",
-    "prettier": "^3.3.3",
+    "prettier": "^3.5.3",
     "rimraf": "^6.0.1",
-    "turbo": "^2.2.3",
-    "typescript": "^5.5.4"
+    "turbo": "^2.5.0",
+    "typescript": "^5.8.3"
   },
-  "packageManager": "pnpm@9.12.2",
+  "packageManager": "pnpm@10.9.0",
   "name": "docs",
   "dependencies": {
-    "dotenv-cli": "^7.4.2"
+    "dotenv-cli": "^8.0.0"
   },
   "pnpm": {
     "patchedDependencies": {
