@@ -10,37 +10,37 @@
     "generate:component": "turbo gen react-component"
   },
   "devDependencies": {
-    "@types/node": "^22.10.7",
+    "@types/node": "^22.14.1",
     "@types/react": "^18.3.12",
     "@types/react-dom": "^18.3.1",
-    "dotenv": "^16.4.5",
-    "eslint": "^8.57.1",
+    "dotenv": "^16.5.0",
+    "eslint": "^9.25.1",
     "eslint-config-xyflow": "workspace:*",
     "react": "^18.3.1",
-    "typescript": "^5.5.4",
+    "typescript": "^5.8.3",
     "xy-tsconfig": "workspace:*"
   },
   "dependencies": {
-    "@codesandbox/sandpack-react": "^2.19.9",
-    "@docsearch/css": "^3.6.2",
-    "@docsearch/react": "^3.6.2",
-    "@heroicons/react": "^2.1.5",
-    "@notionhq/client": "^2.2.15",
-    "@radix-ui/react-accordion": "^1.2.1",
-    "@radix-ui/react-alert-dialog": "^1.1.2",
-    "@radix-ui/react-checkbox": "^1.1.2",
-    "@radix-ui/react-label": "^2.1.0",
-    "@radix-ui/react-popover": "^1.1.2",
-    "@radix-ui/react-radio-group": "^1.2.1",
-    "@radix-ui/react-select": "^2.1.2",
-    "@radix-ui/react-slider": "^1.2.1",
-    "@radix-ui/react-slot": "^1.1.0",
-    "@radix-ui/react-tabs": "^1.1.1",
-    "@react-three/fiber": "9.0.0-rc.7",
+    "@codesandbox/sandpack-react": "^2.20.0",
+    "@docsearch/css": "^3.9.0",
+    "@docsearch/react": "^3.9.0",
+    "@heroicons/react": "^2.2.0",
+    "@notionhq/client": "^2.3.0",
+    "@radix-ui/react-accordion": "^1.2.7",
+    "@radix-ui/react-alert-dialog": "^1.1.10",
+    "@radix-ui/react-checkbox": "^1.2.2",
+    "@radix-ui/react-label": "^2.1.4",
+    "@radix-ui/react-popover": "^1.1.10",
+    "@radix-ui/react-radio-group": "^1.3.3",
+    "@radix-ui/react-select": "^2.2.2",
+    "@radix-ui/react-slider": "^1.3.2",
+    "@radix-ui/react-slot": "^1.2.0",
+    "@radix-ui/react-tabs": "^1.1.8",
+    "@react-three/fiber": "9.1.2",
     "@sindresorhus/slugify": "^2.2.1",
     "@stackblitz/sdk": "^1.11.0",
     "@vercel/blob": "^0.27.1",
-    "@xyflow/react": "^12.5.6",
+    "@xyflow/react": "^12.6.0",
     "@xyflow/xy-ui": "workspace:*",
     "clsx": "^2.1.1",
     "fathom-client": "^3.7.2",
@@ -49,7 +49,7 @@
     "react-player": "^2.16.0",
     "tailwindcss": "^3.4.14",
     "tailwindcss-animate": "^1.0.7",
-    "timeago-react": "^3.0.6",
+    "timeago-react": "^3.0.7",
     "use-error-boundary": "^2.0.6"
   },
   "peerDependencies": {
