@@ -2,6 +2,7 @@
   "name": "reactflow-website",
   "version": "1.0.0",
   "private": true,
+  "type": "module",
   "scripts": {
     "dev": "next --port 3002",
     "build": "next build",
@@ -15,35 +16,36 @@
     "types:check": "tsc --noEmit"
   },
   "dependencies": {
-    "@heroicons/react": "^2.1.5",
-    "@typeform/embed-react": "^3.9.0",
+    "@heroicons/react": "^2.2.0",
+    "@typeform/embed-react": "^4.4.0",
     "@xyflow/react": "^12.6.0",
-    "@xyflow/system": "^0.0.53",
+    "@xyflow/system": "^0.0.57",
     "@xyflow/xy-ui": "workspace:*",
     "d3": "^7.9.0",
     "fast-glob": "^3.3.3",
-    "next": "^15.2.4",
+    "next": "^15.3.1",
     "next-sitemap": "^4.2.3",
     "nextra": "^4.3.0-alpha.19",
     "nextra-theme-docs": "^4.3.0-alpha.19",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
-    "react-icons": "^4.10.1",
-    "timeago-react": "^3.0.6",
+    "react-icons": "^5.5.0",
+    "timeago-react": "^3.0.7",
     "ui-components": "workspace:*",
     "xy-shared": "workspace:*"
   },
   "devDependencies": {
-    "@notionhq/client": "^2.2.15",
+    "@eslint/eslintrc": "^3.3.1",
+    "@notionhq/client": "^2.3.0",
     "@types/d3": "^7.4.3",
-    "@types/node": "^20.17.2",
+    "@types/node": "^22.14.1",
     "@types/react": "^18.3.12",
     "eslint-config-xyflow": "workspace:*",
     "postcss-import": "^16.1.0",
     "postcss-lightningcss": "^1.0.1",
     "source-map-support": "0.5.21",
     "tailwindcss": "^3.4.17",
-    "typescript": "^5.5.4",
+    "typescript": "^5.7.3",
     "xy-tailwind-config": "workspace:*",
     "xy-tsconfig": "workspace:*"
   }
