@@ -14,11 +14,11 @@
     "check-links": "node scripts/check-mdx-links.mjs"
   },
   "dependencies": {
-    "@heroicons/react": "^2.1.5",
+    "@heroicons/react": "^2.2.0",
     "@sindresorhus/slugify": "^2.2.1",
     "@xyflow/react": "12.6.0",
     "@xyflow/xy-ui": "workspace:*",
-    "fast-xml-parser": "^4.5.0",
+    "fast-xml-parser": "^5.2.0",
     "next": "^15.1.7",
     "next-sitemap": "^4.2.3",
     "nextra": "^4.3.0-alpha.19",
@@ -28,14 +28,14 @@
     "xy-shared": "workspace:*"
   },
   "devDependencies": {
-    "@notionhq/client": "^2.2.15",
-    "@types/node": "^20.3.1",
+    "@notionhq/client": "^2.3.0",
+    "@types/node": "^22.14.1",
     "@types/react": "^18.3.12",
     "eslint-config-xyflow": "workspace:*",
     "postcss-import": "^16.1.0",
     "postcss-lightningcss": "^1.0.1",
     "tailwindcss": "^3.4.14",
-    "typescript": "^5.5.4",
+    "typescript": "^5.8.3",
     "xy-tailwind-config": "workspace:*",
     "xy-tsconfig": "workspace:*"
   }
